+++ src/server.ts
@@ -3,6 +3,7 @@
  * @mergeModuleWith <project>
  */
 
+/// <reference lib="webworker" />
 import { type } from "arktype"
 import { createLogger, type LogLevel } from "./log.js"
 import {
@@ -18,6 +19,19 @@ import {
 } from "./types.js"
 import { findTransferables } from "./utils.js"
 
+class MockedWorkerGlobalScope {
+  constructor() {}
+}
+
+const SharedWorkerGlobalScope =
+  globalThis.SharedWorkerGlobalScope ?? MockedWorkerGlobalScope
+
+const DedicatedWorkerGlobalScope =
+  globalThis.DedicatedWorkerGlobalScope ?? MockedWorkerGlobalScope
+
+const ServiceWorkerGlobalScope =
+  globalThis.ServiceWorkerGlobalScope ?? MockedWorkerGlobalScope
+
 /**
  * The sw&rpc server instance, which provides methods to register {@link ProcedureImplementation | procedure implementations},
  * and listens for incoming messages that call those procedures
@@ -25,7 +39,7 @@ import { findTransferables } from "./utils.js"
 export type SwarpcServer<Procedures extends ProceduresMap> = {
   [zProcedures]: Procedures
   [zImplementations]: ImplementationsMap<Procedures>
-  start(self: Window | Worker): void
+  start(): Promise<void>
 } & {
   [F in keyof Procedures]: (
     impl: ProcedureImplementation<
@@ -43,7 +57,9 @@ const abortedRequests = new Set<string>()
  * Creates a sw&rpc server instance.
  * @param procedures procedures the server will implement, see {@link ProceduresMap}
  * @param options various options
- * @param options.worker if provided, the server will use this worker to post messages, instead of sending it to all clients
+ * @param options.worker The worker scope to use, defaults to the `self` of the file where Server() is called.
+ * @param options.loglevel Maximum log level to use, defaults to "debug" (shows everything). "info" will not show debug messages, "warn" will only show warnings and errors, "error" will only show errors.
+ * @param options._scopeType @internal Don't touch, this is used in testing environments because the mock is subpar. Manually overrides worker scope type detection.
  * @returns a SwarpcServer instance. Each property of the procedures map will be a method, that accepts a function implementing the procedure (see {@link ProcedureImplementation}). There is also .start(), to be called after implementing all procedures.
  *
  * An example of defining a server:
@@ -51,9 +67,41 @@ const abortedRequests = new Set<string>()
  */
 export function Server<Procedures extends ProceduresMap>(
   procedures: Procedures,
-  { worker, loglevel = "debug" }: { worker?: Worker; loglevel?: LogLevel } = {}
+  {
+    loglevel = "debug",
+    scope,
+    _scopeType,
+  }: {
+    scope?: WorkerGlobalScope
+    loglevel?: LogLevel
+    _scopeType?: "dedicated" | "shared" | "service"
+  } = {}
 ): SwarpcServer<Procedures> {
   const l = createLogger("server", loglevel)
+
+  // If scope is not provided, use the global scope
+  // This function is meant to be used in a worker, so `self` is a WorkerGlobalScope
+  scope ??= self as WorkerGlobalScope
+
+  function scopeIsShared(
+    scope: WorkerGlobalScope
+  ): scope is SharedWorkerGlobalScope {
+    return scope instanceof SharedWorkerGlobalScope || _scopeType === "shared"
+  }
+
+  function scopeIsDedicated(
+    scope: WorkerGlobalScope
+  ): scope is DedicatedWorkerGlobalScope {
+    return (
+      scope instanceof DedicatedWorkerGlobalScope || _scopeType === "dedicated"
+    )
+  }
+
+  function scopeIsService(
+    scope: WorkerGlobalScope
+  ): scope is ServiceWorkerGlobalScope {
+    return scope instanceof ServiceWorkerGlobalScope || _scopeType === "service"
+  }
 
   // Initialize the instance.
   // Procedures and implementations are stored on properties with symbol keys,
@@ -61,7 +109,7 @@ export function Server<Procedures extends ProceduresMap>(
   const instance = {
     [zProcedures]: procedures,
     [zImplementations]: {} as ImplementationsMap<Procedures>,
-    start: (self: Window) => {},
+    start: async () => {},
   } as SwarpcServer<Procedures>
 
   // Set all implementation-setter methods
@@ -85,7 +133,16 @@ export function Server<Procedures extends ProceduresMap>(
     }) as SwarpcServer<Procedures>[typeof functionName]
   }
 
-  instance.start = (self: Window) => {
+  instance.start = async () => {
+    const port = await new Promise<MessagePort | undefined>((resolve) => {
+      if (!scopeIsShared(scope)) return resolve(undefined)
+      console.log("Awaiting shared worker connection...")
+      scope.addEventListener("connect", ({ ports: [port] }) => {
+        console.log("Shared worker connected with port", port)
+        resolve(port)
+      })
+    })
+
     // Used to post messages back to the client
     const postMessage = async (
       autotransfer: boolean,
@@ -93,17 +150,20 @@ export function Server<Procedures extends ProceduresMap>(
     ) => {
       const transfer = autotransfer ? [] : findTransferables(data)
 
-      if (worker) {
-        self.postMessage(data, { transfer })
-      } else {
-        await (self as any).clients.matchAll().then((clients: any[]) => {
+      if (port) {
+        port.postMessage(data, { transfer })
+      } else if (scopeIsDedicated(scope)) {
+        scope.postMessage(data, { transfer })
+      } else if (scopeIsService(scope)) {
+        await scope.clients.matchAll().then((clients) => {
           clients.forEach((client) => client.postMessage(data, { transfer }))
         })
       }
     }
 
-    // Listen for messages from the client
-    self.addEventListener("message", async (event: MessageEvent) => {
+    const listener = async (
+      event: MessageEvent<any> | ExtendableMessageEvent
+    ): Promise<void> => {
       // Decode the payload
       const { requestId, functionName } = PayloadHeaderSchema(
         type.enumerated(...Object.keys(procedures))
@@ -206,7 +266,21 @@ export function Server<Procedures extends ProceduresMap>(
       } finally {
         abortedRequests.delete(requestId)
       }
-    })
+    }
+
+    // Listen for messages from the client
+    if (scopeIsShared(scope)) {
+      if (!port) throw new Error("SharedWorker port not initialized")
+      console.log("Listening for shared worker messages on port", port)
+      port.addEventListener("message", listener)
+      port.start()
+    } else if (scopeIsDedicated(scope)) {
+      scope.addEventListener("message", listener)
+    } else if (scopeIsService(scope)) {
+      scope.addEventListener("message", listener)
+    } else {
+      throw new Error(`Unsupported worker scope ${scope}`)
+    }
   }
 
   return instance